feat: add `<header>` element for a11y
