+++ src/app/header/header.component.html
@@ -1,4 +1,9 @@
-<app-navigation-tabs [style.flex-grow]="1"></app-navigation-tabs>
-<app-toolbar>
-  <app-light-dark-toggle></app-light-dark-toggle>
-</app-toolbar>
+<header>
+  <app-navigation-tabs
+    [items]="_navigationItems"
+    [style.flex-grow]="1"
+  ></app-navigation-tabs>
+  <app-toolbar>
+    <app-light-dark-toggle></app-light-dark-toggle>
+  </app-toolbar>
+</header>
