+++ src/app/header/header.component.ts
@@ -2,6 +2,7 @@ import { Component } from '@angular/core'
 import { NavigationTabsComponent } from './navigation-tabs/navigation-tabs.component'
 import { ToolbarComponent } from './toolbar/toolbar.component'
 import { LightDarkToggleComponent } from './light-dark-toggle/light-dark-toggle.component'
+import { NAVIGATION_ITEMS } from './navigation-items'
 
 @Component({
   selector: 'app-header',
@@ -14,4 +15,6 @@ import { LightDarkToggleComponent } from './light-dark-toggle/light-dark-toggle.
     LightDarkToggleComponent,
   ],
 })
-export class HeaderComponent {}
+export class HeaderComponent {
+  protected readonly _navigationItems = NAVIGATION_ITEMS
+}
